Moved version declarations of libraries and plugins to properties.
I moved all the version declarations of libraries and plugins to the properties field and use variables in the pom body. Thus, we have a central point of controlling the version numbers that might be related and used at different points in the pom file, e.g. in the dependencies area and the plugin area.